Revert iris.util and put lazy ops in iris._lazy_data.
